cleanup test files vulcan:core
