+++ packages/vulcan-core/test/index.js
@@ -1,4 +1,5 @@
 import './mutations.test';
 import './resolvers.test';
 import './components.test';
-import './containers.test';
+import './queries.test';
+import './withComponents.test';
